src/test/java/com/example/ludogoriesoft/lukeriaerpapi/services/controllers/ImageControllerTest.java
package com.example.ludogoriesoft.lukeriaerpapi.services.controllers;

import com.example.ludogoriesoft.lukeriaerpapi.controllers.ImageController;
import com.example.ludogoriesoft.lukeriaerpapi.exeptions.ApiExceptionHandler;
import com.example.ludogoriesoft.lukeriaerpapi.services.ImageService;
import org.junit.jupiter.api.Assertions;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.mockito.MockitoAnnotations;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
import org.springframework.boot.test.mock.mockito.MockBean;
import org.springframework.context.annotation.ComponentScan;
import org.springframework.context.annotation.FilterType;
import org.springframework.http.MediaType;
import org.springframework.mock.web.MockMultipartFile;
import org.springframework.test.web.servlet.MockMvc;
import org.springframework.test.web.servlet.MvcResult;
import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
import org.springframework.web.multipart.MultipartFile;

import java.io.ByteArrayInputStream;
import java.io.InputStream;

import static org.mockito.Mockito.when;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;

@AutoConfigureMockMvc(addFilters = false)
@WebMvcTest(value = ImageController.class,
        useDefaultFilters = false,
        includeFilters = {
                @ComponentScan.Filter(
                        type = FilterType.ASSIGNABLE_TYPE,
                        value = ImageController.class),
                @ComponentScan.Filter(
                        type = FilterType.ASSIGNABLE_TYPE,
                        value = ApiExceptionHandler.class
                )
        }
)
class ImageControllerTest {
    @Autowired
    private MockMvc mockMvc;

    @MockBean
    private ImageService imageService;

    @BeforeEach
    public void setup() {
        MockitoAnnotations.openMocks(this);
    }

    @Test
    void testUpdateImage() throws Exception {
        byte[] content = "Sample file content".getBytes();

        String filename = "sample.txt";
        InputStream inputStream = new ByteArrayInputStream(content);

        MultipartFile multipartFile = new MockMultipartFile(filename, filename, "text/plain", inputStream);

        when(imageService.saveImageForPackage(multipartFile)).thenReturn(String.valueOf(multipartFile));

        MvcResult mvcResult = mockMvc.perform(MockMvcRequestBuilders.post("/api/v1/images/uploadImageForPackage")
                        .contentType(MediaType.APPLICATION_JSON))
                .andExpect(status().isOk())
                .andReturn();

        String response = mvcResult.getResponse().getContentAsString();
        Assertions.assertNotNull(response);
    }

    @Test
    void testEditImage() throws Exception {
        byte[] content = "Sample file content".getBytes();

        String filename = "sample.txt";
        InputStream inputStream = new ByteArrayInputStream(content);

        MultipartFile multipartFile = new MockMultipartFile(filename, filename, "text/plain", inputStream);


        when(imageService.editImageForPackage(multipartFile, 1L)).thenReturn(String.valueOf(multipartFile));

        MvcResult mvcResult = mockMvc.perform(MockMvcRequestBuilders.post("/api/v1/images/editImageForPackage")
                        .contentType(MediaType.APPLICATION_JSON))
                .andExpect(status().isOk())
                .andReturn();

        String response = mvcResult.getResponse().getContentAsString();
        Assertions.assertNotNull(response);
    }
}
